+++ __tests__/data/campaign_manager.json
@@ -54,16 +54,11 @@
           "floodlightConfigurationId": "213123123",
           "ordinal": "string",
           "floodlightActivityId": "456543345245",
-          "mobileDeviceId": "string",
           "value": "756",
           "encryptedUserIdCandidates": [
             "dfghjbnm"
           ],
           "quantity": "455678",
-          "gclid": "string",
-          "matchId": "string",
-          "dclid": "string",
-          "impressionId": "string",
           "encryptionSource": "AD_SERVING",
           "encryptionEntityId": "3564523",
           "encryptionEntityType": "DCM_ACCOUNT",
@@ -123,16 +118,11 @@
               "ordinal": "string",
               "timestampMicros": "1668624722000000",
               "floodlightActivityId": "456543345245",
-              "mobileDeviceId": "string",
               "quantity": "455678",
               "value": 756,
               "encryptedUserIdCandidates": [
                 "dfghjbnm"
               ],
-              "gclid": "string",
-              "matchId": "string",
-              "dclid": "string",
-              "impressionId": "string",
               "limitAdTracking": false,
               "childDirectedTreatment": false,
               "treatmentForUnderage": false,
@@ -202,16 +192,9 @@
           "floodlightConfigurationId": "213123123",
           "ordinal": "string",
           "floodlightActivityId": "456543345245",
-          "mobileDeviceId": "string",
           "value": "756",
-          "encryptedUserIdCandidates": [
-            "dfghjbnm"
-          ],
           "quantity": "455678",
           "gclid": "string",
-          "matchId": "string",
-          "dclid": "string",
-          "impressionId": "string",
             "encryptionSource": "AD_SERVING",
             "encryptionEntityId": "3564523",
             "encryptionEntityType": "DCM_ACCOUNT",
@@ -259,28 +242,15 @@
       "body": {
         "JSON": {
           "kind": "dfareporting#conversionsBatchUpdateRequest",
-          "encryptionInfo": {
-            "kind": "dfareporting#encryptionInfo",
-            "encryptionSource": "AD_SERVING",
-            "encryptionEntityId": "3564523",
-            "encryptionEntityType": "DCM_ACCOUNT"
-          },
           "conversions": [
             {
               "floodlightConfigurationId": "213123123",
               "ordinal": "string",
               "timestampMicros": "1668624722000000",
               "floodlightActivityId": "456543345245",
-              "mobileDeviceId": "string",
               "quantity": "455678",
               "value": 756,
-              "encryptedUserIdCandidates": [
-                "dfghjbnm"
-              ],
               "gclid": "string",
-              "matchId": "string",
-              "dclid": "string",
-              "impressionId": "string",
               "treatmentForUnderage": false,
               "nonPersonalizedAd": false
             }
