+++ __tests__/data/campaign_manager_router_input.json
@@ -69,20 +69,15 @@
         "ordinal": "string",
         "quantity": "455678",
         "floodlightActivityId": "456543345245",
-        "mobileDeviceId": "string",
         "value": 7,
         "encryptedUserIdCandidates": [
           "dfghjbnm"
         ],
-        "gclid": "string",
-        "matchId": "string",
-        "dclid": "string",
-        "impressionId": "string",
         "limitAdTracking": true,
         "childDirectedTreatment": true,
-          "encryptionSource": "AD_SERVING",
-          "encryptionEntityId": "3564523",
-          "encryptionEntityType": "DCM_ACCOUNT",
+        "encryptionSource": "AD_SERVING",
+        "encryptionEntityId": "3564523",
+        "encryptionEntityType": "DCM_ACCOUNT",
         "requestType": "batchinsert"
       },
       "type": "track",
@@ -164,16 +159,11 @@
         "floodlightConfigurationId": "213123123",
         "ordinal": "string",
         "floodlightActivityId": "456543345245",
-        "mobileDeviceId": "string",
         "quantity": "455678",
         "value": 7,
         "encryptedUserIdCandidates": [
           "dfghjbnm"
         ],
-        "gclid": "string",
-        "matchId": "string",
-        "dclid": "string",
-        "impressionId": "string",
         "limitAdTracking": true,
         "childDirectedTreatment": true,
           "encryptionSource": "AD_SERVING",
